move some types around, cleanup
