+++ modules/headways/charts/HeadwaysSingleChart.tsx
@@ -1,20 +1,11 @@
-import type { UseQueryResult } from '@tanstack/react-query';
 import React, { useMemo } from 'react';
 import { SingleDayLineChart } from '../../../common/components/charts/SingleDayLineChart';
-import type { Station } from '../../../common/types/stations';
+import type { HeadwaysChartProps } from '../../../common/types/charts';
 import { useDelimitatedRoute } from '../../../common/utils/router';
 import { locationDetails } from '../../../common/utils/stations';
-import type { SingleDayDataPoint } from '../../../src/charts/types';
 import { BenchmarkFieldKeys, PointFieldKeys, MetricFieldKeys } from '../../../src/charts/types';
 
-interface HeadwaysSingleChartProps {
-  headways: UseQueryResult<SingleDayDataPoint[]>;
-  toStation: Station | undefined;
-  fromStation: Station | undefined;
-  showLegend?: boolean;
-}
-
-export const HeadwaysSingleChart: React.FC<HeadwaysSingleChartProps> = ({
+export const HeadwaysSingleChart: React.FC<HeadwaysChartProps> = ({
   headways,
   toStation,
   fromStation,
